Type fix.

+++ main/gps.go
@@ -1862,7 +1862,7 @@ func makeAHRSGDL90Report() {
 		if isAHRSInvalidValue(mySituation.AHRSGyroHeading) {
 			hdg = roundToInt16(mySituation.AHRSGyroHeading * 10) // TODO westphae: switch to AHRSMagHeading?
 		} else {
-			hdg = roundToInt16(mySituation.GPSTrueCourse)
+			hdg = roundToInt16(float64(mySituation.GPSTrueCourse))
 		}
 		if !isAHRSInvalidValue(mySituation.AHRSSlipSkid) {
 			slip_skid = roundToInt16(-mySituation.AHRSSlipSkid * 10)
